backend/pyproject.toml
[tool.poetry]
name = "api-compflow"
version = "0.1.0"
description = "API para o trabalho final da disciplina de Engenharia de Software"
authors = ["Mateus Cavalcanti <mateuscavalcanti1607@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.3"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
